// src/models/management/core/operation/operationCostingDetail.interface.ts
interface BaseCostingDetail {
  specialRequest?: string;
  remark?: string;
}

export interface AirCostingDetail extends BaseCostingDetail {
  classOfService?: string;
  tripType?: "ONEWAY" | "ROUNDTRIP" | "MULTICITY";
  adult?: number;
  child?: number;
  infant?: number;
  fullItinerary?: string;
  departureDate?: string;
  arrivalDate?: string;
}

export interface HotelCostingDetail extends BaseCostingDetail {
  roomType?: RoomingType;
  quantity?: number;
  checkIn?: string;
  checkOut?: string;
}

export interface TransportCostingDetail extends BaseCostingDetail {
  carType?: string;
  carModel?: string;
  quantity?: number;
  pickUpDate?: string;
  pickUpLocation?: string;
  dropOffLocation?: string;
}

export interface GuideCostingDetail extends BaseCostingDetail {
  guideType?: "DOMESTIC" | "OUTBOUND" | "INBOUND";
  quantity?: number;
  destination?: string;
}

export interface VisaCostingDetail extends BaseCostingDetail {}